v3.2.8

+++ activerecord/lib/active_record/relation.rb
@@ -77,6 +77,7 @@ module ActiveRecord
     end
 
     def initialize_copy(other)
+      @bind_values = @bind_values.dup
       reset
     end
 
@@ -402,6 +403,8 @@ module ActiveRecord
     # If you need to destroy dependent associations or call your <tt>before_*</tt> or
     # +after_destroy+ callbacks, use the +destroy_all+ method instead.
     def delete_all(conditions = nil)
+      raise ActiveRecordError.new("delete_all doesn't support limit scope") if self.limit_value
+
       IdentityMap.repository[symbolized_base_class] = {} if IdentityMap.enabled?
       if conditions
         where(conditions).delete_all
@@ -453,7 +456,7 @@ module ActiveRecord
     end
 
     def to_sql
-      @to_sql ||= klass.connection.to_sql(arel)
+      @to_sql ||= klass.connection.to_sql(arel, @bind_values.dup)
     end
 
     def where_values_hash
